style: drop a redundant int cast Rector flagged
Pushed the previous commit before checking the Rector gate. The cast on
array_key_first was redundant and the dry-run caught it.

+++ app/Services/Extraction/KeyValueExtractor.php
@@ -52,7 +52,7 @@ class KeyValueExtractor
             return null;
         }
 
-        $firstIndex = (int) array_key_first($candidates);
+        $firstIndex = array_key_first($candidates);
 
         return $value + ['signals' => new FieldExtractionSignals(
             true,
